InventoryManagementSystem: Feat ProductRepository

+++ playground-inventory-management-system/src/main/java/com/currenjin/inventory/infrastructure/InventoryJPARepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.currenjin.inventory.domain.Inventory;
 import com.currenjin.inventory.domain.repository.InventoryRepository;
 
-public interface InventoryJPARepository extends InventoryRepository, JpaRepository<Inventory, Integer> {
+public interface InventoryJPARepository extends InventoryRepository, JpaRepository<Inventory, Long> {
 }
